Change index.html structure and minor styling
Changed the website design after realizing I could
use better nested containers for the landing page
sections.
Added the class 'container' to all the elements
that are going to stay in the central column.
Added a logo image that is also the website icon.
Added a root with color palette variables after
learning about variables calling in CSS.
Styled the header.