Adjust styling and edit index.html
Added custom descriptions to cards in second section
as well as fixed alt message of images for screen
readers.
Fixed CSS style for the second section to accommodate
changes made in html.
Minor tweaking to the page's style, added a box-shadow
to the hero image in first section.
Note to self: kept right and left containers on fourth
section to pick up project in the future.