Fix responsive problems
Setting the ground to fix some responsive problems
by adding flex-wrap to card in the second section.
Edited third exercise (dips) description to fit in
the same amount of lines as the others.
Set a width of 100% to hero image on first section
to fix responsive problems (width was never set so
flex-basis was keeping the img default size).